Olken.csv model analysis
The dataset includes the variables municipality, %missing, and audit. audit is a binary variable taking
0 if a municipality did not conduct a corruption audit, and 1 otherwise and %missing is the proportion of
funds that went missing (so it falls between 0 and 1)
. Municipal councils decide whether or not to conduct an
audit. About 30% of observations for %missing are missing. Your friend used OLS to analyze lm(%missing
→
audit). Theyﬁndω1 is equal to
↑
.15, which they claim is the causal effect of audits on missing funds.
a) Do you ﬁnd the random sampling assumption for this dataset to be plausible? Why or why not?
b) Do you ﬁnd the zero conditional mean assumption for this model to be plausible? Why or why not?
c) Given your answers, do you believe that the municipal audits in Argentina lowered missing funds?